// src/types/Countries.ts
export interface CountriesTS {
  name: string;
  translations: {
    pt: string;
    br: string;
    de: string;
    es: string;
    fa: string;
    fr: string;
    hr: string;
    hu: string;
    it: string;
    ja: string;
    nl: string;
  };
  capital: string;
  population: number;
  region: string;
  numericCode: string;
  flags: {
    png: string;
  };
}